Extracted Customer and CustomerInitiative to its own files, renamed Customers to CustomerTracker #22
